requirements.txt
flask
pyinstaller
pywebview
yfinance
pandas
numpy
